feat(denuncias): adicionar botao animado e modularizar estilos
