+++ components/complaints/error-state.jsx
@@ -1,5 +1,5 @@
 import { Pressable, Text, View } from 'react-native';
-import { complaintsStyles } from '@/styles/complaints.styles';
+import { complaintsStyles } from '@/styles/complaints';
 import { useColorScheme } from '@/hooks/useColorScheme';
 
 export function ErrorState({ message, onRetry }) {
