+++ components/complaints/loading-state.jsx
@@ -1,5 +1,5 @@
 import { ActivityIndicator, Text, View } from 'react-native';
-import { complaintsStyles } from '@/styles/complaints.styles';
+import { complaintsStyles } from '@/styles/complaints';
 import { useColorScheme } from '@/hooks/useColorScheme';
 
 export function LoadingState() {
