+++ components/filter-chips/filter-chips.jsx
@@ -1,11 +1,18 @@
 import { FILTER_CHIPS } from '@/constants/complaints.costants';
 import { filterChipsStyles } from '@/styles/filter-chips.style';
 import { useColorScheme } from '@/hooks/useColorScheme';
+import { useHaptics } from '@/hooks/useHaptics';
 import { ScrollView, Text, TouchableOpacity } from 'react-native';
 
 export function FilterChips({ activeChip = null, onChipPress }) {
   const colorScheme = useColorScheme();
   const styles = filterChipsStyles(colorScheme);
+  const { triggerHaptics } = useHaptics();
+
+  const handleChipPress = (value) => {
+    triggerHaptics('soft');
+    onChipPress(value);
+  };
 
   return (
     <ScrollView
@@ -18,7 +25,7 @@ export function FilterChips({ activeChip = null, onChipPress }) {
         <TouchableOpacity
           key={chip.label}
           style={[styles.chip, activeChip === chip.value ? styles.chipActive : null]}
-          onPress={() => onChipPress(chip.value)}
+          onPress={() => handleChipPress(chip.value)}
         >
           <Text style={[styles.chipText, activeChip === chip.value ? styles.chipTextActive : null]}>
             {chip.label}
