+++ styles/complaints/list.styles.jsx
@@ -1,0 +1,11 @@
+import { StyleSheet } from 'react-native';
+
+export const listStyles = (colorScheme) => {
+  return StyleSheet.create({
+    listContent: {
+      padding: 14,
+      gap: 10,
+      paddingBottom: 80,
+    },
+  });
+};
