+++ styles/mapScreen/container.styles.jsx
@@ -1,0 +1,13 @@
+import { StyleSheet } from 'react-native';
+
+export const containerStyles = (colorScheme) => {
+  return StyleSheet.create({
+    container: { 
+      flex: 1 
+    },
+    
+    map: { 
+      flex: 1 
+    },
+  });
+};
